Fix minOrderValueGp not appearing in the config panel - RuneLite can't render long
RuneLite's config UI only knows how to render int/boolean/String/double/enum/Color
config item types, not long - a config method returning long silently gets no
field at all in the settings panel, no error either. minOrderValueGp was declared
as long (its sibling maxGpPerOrder right next to it is int, which is why that one
rendered fine and this one didn't). Changed it to int - order.totalValue() is a
long, so the comparison still widens correctly, and int's ~2.1 billion range is
far beyond any real minimum-order-value setting.

+++ plugins/ppo-flipper-star/src/main/java/net/runelite/client/plugins/microbot/ppoflipperstar/Guardrails.java
@@ -121,7 +121,7 @@ public class Guardrails {
         // a GE slot, a fill wait, and the attention to notice/react to it are worth more than the
         // gp involved either way - not worth doing at all regardless of direction. Applies to
         // both BUY and SELL for the same reason (a trivially small BUY wastes a slot identically).
-        long minOrderValue = config.minOrderValueGp();
+        int minOrderValue = config.minOrderValueGp();
         if (minOrderValue > 0 && order.totalValue() < minOrderValue) {
             return String.format(
                 "total order value %d gp is below the minimum worth trading (%d gp)",
